Add custom_claims argument for encoding access tokens
This is where users can pass in arbitrary data, such as ip address,
username, object/database ids, or anything else that can be json
serializable, and have it be stored in the token. I still need to
provide some helper functions for accessing the data (need to store it
in flask.g, and probably have an easier way of getting it, such as a
method like 'get_custom_claim(key, default)').