Remove token last used field in the token store
This would be a nice feature to have, but it can lead lead to (very
rare) race conditions. simplekv has no way to support atomic retreive
and replace. We could revoke a key after grabbing the key to update,
then overwrite the revoke status. This also makes storing an effecient
mapping of jtis that exist to an identity hard.

Currently, I'm thinking if a user wants this, they should add it as part
of their application logic, something like

@app.route('foobar')
@jwt_refresh_token_required
def refresh:
    udate_token_last_used(jti)  # Need a way for users to get the jti
    access_token = create_access_token()
    return ...