unset_jwt_cookies now removes csrf tokesn too
Practically, there isn't any security concerns by leaving them set. We
don't do any verification on these cookies when they are sent to a
protected endpoint, and if we generated new tokens the values in those
cookies would be updated. This is just to make sure we are cleaning up
after ourselfs (refs #34)