Move chat location from irc to gitter. (#196)
IRC is great, but it's significatnly harder to have a history on for
multiple people unless they are all running their own IRC bouncers.
Because I am working on getting more maintainers for this project, I
have dediced to move it over to gitter for ease of use. I will still be
hanging out on IRC though.