Add option to include `current_user` in jina templates
This can be turned on via:

```python
JWTManager(app, add_context_processor=True)
```

or

```python
jwt.init_app(app, add_context_processor=True)
```

I did not make turning this on the default because it could be a
breaking change. Come the next major release I will likely change this
to be on by default.

Fixes #478